requirements.txt
Package            Version
------------------ ----------
asgiref            3.8.1
certifi            2024.8.30
cffi               1.17.1
charset-normalizer 3.4.0
Django             5.1.3
ffmpeg-python      0.2.0
future             1.0.0
idna               3.10
pip                22.0.2
pycparser          2.22
pydub              0.25.1
requests           2.32.3
setuptools         59.6.0
sqlparse           0.5.2
srt                3.5.3
tqdm               4.67.0
typing_extensions  4.12.2
urllib3            2.2.3
vosk               0.3.45
websockets         14.1
youtube-dl         2021.12.17
yt-dlp             2024.11.4
